BitcoinWalletProvider: abstract out duplication from ledger and js providers
