+++ packages/bitcoin-js-wallet-provider/package.json
@@ -21,7 +21,8 @@
     "@babel/runtime": "^7.4.3",
     "@liquality/bitcoin-utils": "^0.4.31",
     "@liquality/jsonrpc-provider": "^0.4.31",
-    "@liquality/provider": "^0.4.31",
+    "@liquality/bitcoin-wallet-provider": "^0.4.31",
+    "@liquality/wallet-provider": "^0.4.31",
     "@liquality/utils": "^0.4.31",
     "bignumber.js": "^9.0.0",
     "bip32": "^1.0.2",
