+++ packages/bitcoin-ledger-provider/lib/BitcoinLedgerProvider.js
@@ -3,6 +3,7 @@ import bip32 from 'bip32'
 import * as bitcoin from 'bitcoinjs-lib'
 
 import LedgerProvider from '@liquality/ledger-provider'
+import BitcoinWalletProvider from '@liquality/bitcoin-wallet-provider'
 import HwAppBitcoin from '@ledgerhq/hw-app-btc'
 
 import {
@@ -10,36 +11,16 @@ import {
 } from '@liquality/crypto'
 import {
   compressPubKey,
-  getAddressNetwork,
-  AddressTypes,
-  selectCoins
+  getAddressNetwork
 } from '@liquality/bitcoin-utils'
 import networks from '@liquality/bitcoin-networks'
-import { Address, addressToString } from '@liquality/utils'
+import { addressToString } from '@liquality/utils'
 
 import { version } from '../package.json'
 
-const ADDRESS_GAP = 20
-const NONCHANGE_ADDRESS = 0
-const CHANGE_ADDRESS = 1
-const NONCHANGE_OR_CHANGE_ADDRESS = 2
-
-const ADDRESS_TYPE_TO_LEDGER_PREFIX = {
-  'legacy': 44,
-  'p2sh-segwit': 49,
-  'bech32': 84
-}
-
-export default class BitcoinLedgerProvider extends LedgerProvider {
+export default class BitcoinLedgerProvider extends BitcoinWalletProvider(LedgerProvider) {
   constructor (network = networks.bitcoin, addressType = 'bech32') {
-    if (!AddressTypes.includes(addressType)) {
-      throw new Error(`addressType must be one of ${AddressTypes.join(',')}`)
-    }
-    const derivationPath = `${ADDRESS_TYPE_TO_LEDGER_PREFIX[addressType]}'/${network.coinType}'/0'/`
-    super(HwAppBitcoin, derivationPath, network, 'BTC')
-    this._addressType = addressType
-    this._derivationPath = derivationPath
-    this._network = network
+    super(network, addressType, [HwAppBitcoin, network, 'BTC'])
     this._walletPublicKeyCache = {}
   }
 
@@ -184,89 +165,6 @@ export default class BitcoinLedgerProvider extends LedgerProvider {
     return valueBuffer.reverse()
   }
 
-  async getInputsForAmount (_targets, numAddressPerCall = 100) {
-    let addressIndex = 0
-    let changeAddresses = []
-    let nonChangeAddresses = []
-    let addressCountMap = {
-      change: 0,
-      nonChange: 0
-    }
-
-    const feePerBytePromise = this.getMethod('getFeePerByte')()
-    let feePerByte = false
-
-    while (addressCountMap.change < ADDRESS_GAP || addressCountMap.nonChange < ADDRESS_GAP) {
-      let addrList = []
-
-      if (addressCountMap.change < ADDRESS_GAP) {
-        // Scanning for change addr
-        changeAddresses = await this.getAddresses(addressIndex, numAddressPerCall, true)
-        addrList = addrList.concat(changeAddresses)
-      } else {
-        changeAddresses = []
-      }
-
-      if (addressCountMap.nonChange < ADDRESS_GAP) {
-        // Scanning for non change addr
-        nonChangeAddresses = await this.getAddresses(addressIndex, numAddressPerCall, false)
-        addrList = addrList.concat(nonChangeAddresses)
-      }
-
-      let utxos = await this.getMethod('getUnspentTransactions')(addrList)
-      utxos = utxos.map(utxo => {
-        const addr = addrList.find(a => a.equals(utxo.address))
-        return {
-          ...utxo,
-          value: BigNumber(utxo.amount).times(1e8).toNumber(),
-          derivationPath: addr.derivationPath
-        }
-      })
-
-      const usedAddresses = []
-      // const usedAddresses = confirmedAdd.concat(utxosMempool) // TODO: USED ADDRESSES
-      // utxos = utxos // TODO: Filter out utxos in the mempool that have already been used? Does the node already do this?
-      //   .filter(utxo => utxosMempool.filter(mempoolUtxo => utxo.txid === mempoolUtxo.prevtxid).length === 0)
-
-      if (feePerByte === false) feePerByte = await feePerBytePromise
-      const minRelayFee = await this.getMethod('getMinRelayFee')()
-
-      const targets = _targets.map((target, i) => ({ id: 'main', value: target.value }))
-
-      const { inputs, outputs, fee } = selectCoins(utxos, targets, Math.ceil(feePerByte), minRelayFee)
-
-      if (inputs && outputs) {
-        let change = outputs.find(output => output.id !== 'main')
-
-        if (change && change.length) {
-          change = change[0].value
-        }
-
-        return {
-          inputs,
-          change,
-          fee
-        }
-      }
-
-      for (let address of addrList) {
-        const isUsed = usedAddresses.find(a => address.equals(a))
-        const isChangeAddress = changeAddresses.find(a => address.equals(a))
-        const key = isChangeAddress ? 'change' : 'nonChange'
-
-        if (isUsed) {
-          addressCountMap[key] = 0
-        } else {
-          addressCountMap[key]++
-        }
-      }
-
-      addressIndex += numAddressPerCall
-    }
-
-    throw new Error('Not enough balance')
-  }
-
   async getLedgerInputs (unspentOutputs) {
     const app = await this.getApp()
 
@@ -275,25 +173,6 @@ export default class BitcoinLedgerProvider extends LedgerProvider {
       const tx = app.splitTransaction(hex, true)
       return [ tx, utxo.vout ]
     }))
-  }
-
-  getAddressFromPublicKey (publicKey) {
-    if (this._addressType === 'legacy') {
-      return bitcoin.payments.p2pkh({ pubkey: publicKey, network: this._network }).address
-    } else if (this._addressType === 'p2sh-segwit') {
-      return bitcoin.payments.p2sh({
-        redeem: bitcoin.payments.p2wpkh({ pubkey: publicKey, network: this._network }),
-        network: this._network }).address
-    } else if (this._addressType === 'bech32') {
-      return bitcoin.payments.p2wpkh({ pubkey: publicKey, network: this._network }).address
-    }
-  }
-
-  async _importAddresses () {
-    const change = await this.getAddresses(0, 200, true)
-    const nonChange = await this.getAddresses(0, 200, false)
-    const all = [...nonChange, ...change].map(addressToString)
-    await this.getMethod('importAddresses')(all)
   }
 
   async _getWalletPublicKey (path) {
@@ -312,7 +191,7 @@ export default class BitcoinLedgerProvider extends LedgerProvider {
     return walletPublicKey
   }
 
-  async getLedgerAddresses (startingIndex, numAddresses, change = false) {
+  async baseDerivationNode () {
     const walletPubKey = await this.getWalletPublicKey(this._baseDerivationPath)
     const compressedPubKey = compressPubKey(walletPubKey.publicKey)
     const node = bip32.fromPublicKey(
@@ -320,116 +199,7 @@ export default class BitcoinLedgerProvider extends LedgerProvider {
       Buffer.from(walletPubKey.chainCode, 'hex'),
       this._network
     )
-
-    const addresses = []
-    const lastIndex = startingIndex + numAddresses
-    const changeVal = change ? '1' : '0'
-
-    for (let currentIndex = startingIndex; currentIndex < lastIndex; currentIndex++) {
-      const subPath = changeVal + '/' + currentIndex
-      const publicKey = node.derivePath(subPath).publicKey
-      const address = this.getAddressFromPublicKey(publicKey)
-      const path = this._baseDerivationPath + subPath
-
-      addresses.push(new Address({
-        address,
-        publicKey: publicKey,
-        derivationPath: path,
-        index: currentIndex
-      }))
-    }
-
-    return addresses
-  }
-
-  async _getUsedUnusedAddresses (numAddressPerCall = 100, addressType) {
-    const usedAddresses = []
-    const addressCountMap = { change: 0, nonChange: 0 }
-    const unusedAddressMap = { change: null, nonChange: null }
-
-    let addrList
-    let addressIndex = 0
-    let changeAddresses = []
-    let nonChangeAddresses = []
-
-    /* eslint-disable no-unmodified-loop-condition */
-    while (
-      (addressType === NONCHANGE_OR_CHANGE_ADDRESS && (
-        addressCountMap.change < ADDRESS_GAP || addressCountMap.nonChange < ADDRESS_GAP)
-      ) ||
-      (addressType === NONCHANGE_ADDRESS && addressCountMap.nonChange < ADDRESS_GAP) ||
-      (addressType === CHANGE_ADDRESS && addressCountMap.change < ADDRESS_GAP)
-    ) {
-      /* eslint-enable no-unmodified-loop-condition */
-      addrList = []
-
-      if ((addressType === NONCHANGE_OR_CHANGE_ADDRESS || addressType === CHANGE_ADDRESS) &&
-           addressCountMap.change < ADDRESS_GAP) {
-        // Scanning for change addr
-        changeAddresses = await this.getAddresses(addressIndex, numAddressPerCall, true)
-        addrList = addrList.concat(changeAddresses)
-      } else {
-        changeAddresses = []
-      }
-
-      if ((addressType === NONCHANGE_OR_CHANGE_ADDRESS || addressType === NONCHANGE_ADDRESS) &&
-           addressCountMap.nonChange < ADDRESS_GAP) {
-        // Scanning for non change addr
-        nonChangeAddresses = await this.getAddresses(addressIndex, numAddressPerCall, false)
-        addrList = addrList.concat(nonChangeAddresses)
-      }
-
-      const transactionCounts = await this.getMethod('getAddressTransactionCounts')(addrList)
-
-      for (let address of addrList) {
-        const isUsed = transactionCounts[address] > 0
-        const isChangeAddress = changeAddresses.find(a => address.equals(a))
-        const key = isChangeAddress ? 'change' : 'nonChange'
-
-        if (isUsed) {
-          usedAddresses.push(address)
-          addressCountMap[key] = 0
-          unusedAddressMap[key] = null
-        } else {
-          addressCountMap[key]++
-
-          if (!unusedAddressMap[key]) {
-            unusedAddressMap[key] = address
-          }
-        }
-      }
-
-      addressIndex += numAddressPerCall
-    }
-
-    let firstUnusedAddress
-    const indexNonChange = unusedAddressMap.nonChange ? unusedAddressMap.nonChange.index : Infinity
-    const indexChange = unusedAddressMap.change ? unusedAddressMap.change.index : Infinity
-
-    if (indexNonChange <= indexChange) firstUnusedAddress = unusedAddressMap.nonChange
-    else firstUnusedAddress = unusedAddressMap.change
-
-    return {
-      usedAddresses,
-      unusedAddress: unusedAddressMap,
-      firstUnusedAddress
-    }
-  }
-
-  async getUsedAddresses (numAddressPerCall = 100) {
-    return this._getUsedUnusedAddresses(numAddressPerCall, NONCHANGE_OR_CHANGE_ADDRESS)
-      .then(({ usedAddresses }) => usedAddresses)
-  }
-
-  async getUnusedAddress (change = false, numAddressPerCall = 100) {
-    const addressType = change ? CHANGE_ADDRESS : NONCHANGE_ADDRESS
-    const key = change ? 'change' : 'nonChange'
-    return this._getUsedUnusedAddresses(numAddressPerCall, addressType)
-      .then(({ unusedAddress }) => unusedAddress[key])
-  }
-
-  async getAddresses (startingIndex = 0, numAddresses = 1, change = false) {
-    return this.getLedgerAddresses(startingIndex, numAddresses, change)
+    return node
   }
 
   async getConnectedNetwork () {
@@ -445,8 +215,3 @@ export default class BitcoinLedgerProvider extends LedgerProvider {
 }
 
 BitcoinLedgerProvider.version = version
-BitcoinLedgerProvider.addressType = {
-  NONCHANGE_ADDRESS,
-  CHANGE_ADDRESS,
-  NONCHANGE_OR_CHANGE_ADDRESS
-}
