+++ packages/bitcoin-wallet-provider/lib/index.js
@@ -1,0 +1,3 @@
+import BitcoinWalletProvider from './BitcoinWalletProvider'
+
+export default BitcoinWalletProvider
